Change the meaning option which_states, remove security checks.
Given a system with N states, copying states' result into
output becomes much nicer if only the first n<N states
are kept and all n+1 states are discarded. This is different
to the former implementation where which_states was a vector
enumerating the states to return. As pointed out below, this
implementation was prone to seg fault.

With checks removed, the former code to copy results reads

  for(auto it = extract_states.begin(); it != extract_states.end(); it++) {
    output.at(0,*it) = NV_Ith_S(y,*it - 1);
  }

Now, if the system has 5 states, and user requests states
1,2,5 to be returned, than in turn *it is 1,2,5. However,
output has columns prepared for 3 states only, as
this is the size of which_states. Hence,

  output.at(0, 5)

indexes out-of-range. The current implementation only allows
to discard states at the end of the state vector.

For those reasons, I changed the meaning of which_states.

Removed security checks
-----------------------
The vector y

  y = N_VNew_Serial(neq);

is initialized with size neq. Now we check, if the number of
requested states noutStates is larger than neq and report an
error if so. All indexing in then limited by noutStates which
then is already guarenteed to not exceed the range of y.