Refactor main time loop: as<>() and try{}
Moving as<>() outside of the loop gives a nice
speed-up of factor 5.

try{} now includes the eniter main loop. This
might lead to a speed-up as the try-clause is
not entered anew for each time point. Moreover,
the for-loop governing the integration becomes
try{} free, which _might_ lead to better
comptile time optimizations.