Fix events
cvode signals the presence of an event by the return flag of
cvode(). However, at least if the events conicides with a requested output
time point, as it is the case for us, the flag comes too late. This is,
the main time loop is advanced one step past the event time, when the
event is signaled. The current solution is to decrement the loop counter
by one in case of an event. This is not great, but seems to work.